+++ CHANGELOG.md
@@ -8,13 +8,21 @@
 
 ### Changed
 
-- `.paged` layouts now respect `bounds` for each page. `itemInsets` has been removed; use `bounds.padding` instead.
-
 ### Misc
 
 ### Internal
 
 # Past Releases
+
+# [11.0.0] - 2023-06-15
+
+### Removed
+
+- `.paged` layout's `itemInsets` has been removed; use `bounds.padding` instead.
+
+### Changed
+
+- `.paged` layouts now respect `bounds` for each page.
 
 # [10.3.1] - 2023-06-06
 
@@ -942,7 +950,8 @@ listActions.scrolling.scrollToSection(
 Earlier releases were ad-hoc and not tracked. To see all changes, please reference [closed PRs on Github](https://github.com/kyleve/Listable/pulls?q=is%3Apr+is%3Aclosed).
 
 
-[Main]: https://github.com/square/Listable/square/10.3.1...HEAD
+[Main]: https://github.com/square/Listable/square/11.0.0...HEAD
+[11.0.0]: https://github.com/square/Listable/compare/10.3.1...11.0.0
 [10.3.1]: https://github.com/square/Listable/compare/10.3.0...10.3.1
 [10.3.0]: https://github.com/square/Listable/compare/10.2.0...10.3.0
 [10.2.0]: https://github.com/square/Listable/compare/10.1.0...10.2.0